feat: create note-card and new-note components
It was necessary to create cards to display the notes that the user
created, a
card for notes already written and one for new notes.

This commit creates two components, the note-card which is a card
with the notes already written previously and the new-note-card
component that this is a card where you will write the desired note.